+++ mantis-publish/mantis-publish-core/src/main/java/io/mantisrx/publish/StreamManager.java
@@ -136,19 +136,6 @@ public class StreamManager {
                 .orElse(false);
     }
 
-    private List<String> sanitizeStreamSubjects(final List<String> subjects) {
-        return subjects.stream()
-                .map(s -> {
-                    if (s.toLowerCase().equals("observable") ||
-                            s.toLowerCase().equals("stream")) {
-                        // Translate the legacy default stream names to map to the default stream.
-                        return StreamType.DEFAULT_EVENT_STREAM;
-                    } else {
-                        return s;
-                    }
-                }).collect(Collectors.toList());
-    }
-
     private void handleDuplicateSubscriptionId(final Subscription sub) {
         String subId = sub.getSubscriptionId();
         Optional.ofNullable(subscriptionIdToStreamsMap.get(subId))
@@ -156,7 +143,7 @@ public class StreamManager {
     }
 
     synchronized void addStreamSubscription(final Subscription sub) {
-        List<String> streams = sanitizeStreamSubjects(sub.getSubjects());
+        List<String> streams = sub.getSubjects();
         LOG.info("adding subscription {} with streams {}", sub, streams);
 
         handleDuplicateSubscriptionId(sub);
@@ -224,7 +211,7 @@ public class StreamManager {
 
     synchronized boolean removeStreamSubscription(final Subscription sub) {
         LOG.info("removing subscription {}", sub);
-        final List<String> streams = sanitizeStreamSubjects(sub.getSubjects());
+        final List<String> streams = sub.getSubjects();
         removeSubscriptionId(streams, sub.getSubscriptionId());
         subscriptionIdToStreamsMap.remove(sub.getSubscriptionId());
