+++ mantis-publish/mantis-publish-core/src/main/java/io/mantisrx/publish/api/StreamType.java
@@ -21,7 +21,8 @@ public class StreamType {
     /**
      * Default Stream name for emitting events.
      */
-    public static final String DEFAULT_EVENT_STREAM = "defaultStream";
+    //public static final String DEFAULT_EVENT_STREAM = "defaultStream";
+	public static final String DEFAULT_EVENT_STREAM = "__default__";
 
     /**
      * Stream name for emitting request events.
